+++ RUFAS/data_structures/animal_to_manure_connection.py
@@ -120,6 +120,7 @@ class ManureStream:
         Volume of the manure stream (m^3).
     pen_manure_data : PenManureData | None
        Optional, more specific information about the manure and the pen or pens that produced it.
+
     """
 
     water: float
@@ -165,6 +166,28 @@ class ManureStream:
         )
 
     @property
+    def is_empty(self) -> bool:
+        """
+        Returns True if all nutrient, solids, and volume values are zero
+        and pen_manure_data is None.
+        """
+        return self.pen_manure_data is None and all(
+            value == 0.0
+            for value in [
+                self.water,
+                self.ammoniacal_nitrogen,
+                self.nitrogen,
+                self.phosphorus,
+                self.potassium,
+                self.ash,
+                self.non_degradable_volatile_solids,
+                self.degradable_volatile_solids,
+                self.total_solids,
+                self.volume,
+            ]
+        )
+
+    @property
     def total_volatile_solids(self) -> float:
         """Amount of the total volatile solids (kg)."""
         return self.non_degradable_volatile_solids + self.degradable_volatile_solids
@@ -177,3 +200,20 @@ class ManureStream:
     def clear_pen_manure_data(self) -> None:
         """Clears the pen manure data instance."""
         self.pen_manure_data = None
+
+    @classmethod
+    def make_empty_manure_stream(cls) -> "ManureStream":
+        """Factory method for making empty ManureStreams."""
+        return ManureStream(
+            water=0.0,
+            ammoniacal_nitrogen=0.0,
+            nitrogen=0.0,
+            phosphorus=0.0,
+            potassium=0.0,
+            ash=0.0,
+            non_degradable_volatile_solids=0.0,
+            degradable_volatile_solids=0.0,
+            total_solids=0.0,
+            volume=0.0,
+            pen_manure_data=None,
+        )